+++ safeparse/logging/logger.py
@@ -2,11 +2,13 @@ import logging
 from collections import deque
 from rich import print
 
+from safeparse.setting import EVENT_LOG
+
 logger = logging.getLogger("safeparse")
 logger.setLevel(logging.INFO)  # You can change the level
 
 # File handler: logs to 'event.log'
-file_handler = logging.FileHandler("event.log")
+file_handler = logging.FileHandler(EVENT_LOG)
 file_handler.setFormatter(logging.Formatter("[%(asctime)s] %(levelname)s %(message)s"))
 
 # Avoid duplicated handlers
@@ -15,7 +17,7 @@ if not logger.handlers:
 
 
 def print_log():
-    with open("event.log", "r") as log:
+    with open(EVENT_LOG, "r") as log:
         log_items = deque(log, maxlen=20)
 
     for items in log_items:
